Replaced eslint with Standard for formatting
I replaced the eslint coding style with Standard as I genuinely believe it to be the superior style. Primarily due to the lack of project specific settings, you cannot change the coding style so it's always the same.

This means I updated the package.json to include the correct dependency and added a script called format. Running npm format will use Standard to automatically fix many formatting errors.

I also added Josh and myself to the package.json as contributors.